+++ tests/spike-manifest-scoping.test.cjs
